Reorganized methods
